+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @xnetjs/storage
 
+## 2.5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xnetjs/sqlite@2.5.0
+  - @xnetjs/crypto@2.5.0
+  - @xnetjs/core@2.5.0
+
 ## 2.4.0
 
 ### Patch Changes